separate out MinMaxSelect component
